wip SykmeldingHenterTest
