+++ src/main/resources/db/migration/V8__sykmelding_fjern_person.sql
@@ -1,0 +1,4 @@
+ALTER TABLE SYKMELDING
+    DROP COLUMN PERSON,
+    DROP COLUMN SISTE_SYKMELDINGSTATUS_ID
+;
